seperated components from frames
