--- client/src/Components/Frames/Header/CurrentTime/CurrentTime.scss
@@ -1,12 +1,0 @@
-@use '../../../../sass/partials/variables' as *;
-
-.toolbar {
-	&__time {
-		font-size: 1.5rem;
-		text-shadow: $text-shadow-white;
-	}
-
-	&__day {
-		text-shadow: $text-shadow-white;
-	}
-}
