--- client/src/Components/Frames/Header/HeaderFrame.scss
@@ -1,26 +1,0 @@
-@use '../../../sass/partials/variables' as *;
-
-.HeaderFrame {
-	background-color: rgba($color: grey, $alpha: 0.15);
-	position: fixed;
-	top: 0;
-	width: 100%;
-	height: 8vh;
-	z-index: 0;
-
-	display: flex;
-	justify-content: space-between;
-	align-items: center;
-
-	padding: 0 1rem;
-}
-
-.logo {
-	&__title {
-		text-shadow: $text-shadow-white;
-		cursor: pointer;
-	}
-}
-
-.banner {
-}
