--- client/src/Components/Frames/Main/Chat/Chat.scss
@@ -1,64 +1,0 @@
-@use '../../../../sass/partials/variables' as *;
-
-.chat {
-	width: 400px;
-	height: 600px;
-	background-color: rgba($color: #ccc, $alpha: 0.5);
-	border-radius: 3px;
-
-	display: flex;
-	flex-direction: column;
-	justify-content: space-between;
-
-	&__text {
-		display: flex;
-		flex-direction: column-reverse;
-		padding: 0.5rem;
-		overflow: hidden;
-	}
-
-	&__form {
-		flex-grow: 1;
-		display: flex;
-	}
-
-	&__input {
-		display: flex;
-		padding: 0.5rem;
-		align-items: center;
-	}
-
-	&__input-field {
-		margin-left: 0.5rem;
-		padding: 0.5rem;
-		font-size: 1.2rem;
-		width: 6.5rem;
-		height: 2rem;
-		border: none;
-		border-radius: 3px;
-		background-color: rgba($color: #333, $alpha: 0.5);
-		color: #333;
-
-		&:focus,
-		&:focus-visible {
-			outline: none;
-		}
-
-		&::placeholder {
-			font-size: 1.2rem;
-			font-weight: bold;
-			color: #333;
-			text-shadow: $text-shadow-dark;
-		}
-		flex-grow: 1;
-	}
-}
-
-.button-chat {
-	margin: 0 0 0 0.5rem;
-	background-color: #555;
-
-	&:hover {
-		background-color: rgb(19, 107, 24);
-	}
-}
