--- client/src/Components/Frames/Side/Tasks/Tasks.scss
@@ -1,15 +1,0 @@
-@use '../../../../sass/partials/variables' as *;
-
-.tasks {
-	text-align: right;
-	background-color: rgba(0, 0, 0, 0.2);
-	padding: 1rem;
-	h1 {
-		margin-bottom: 1rem;
-		text-shadow: $text-shadow-white;
-	}
-	ul {
-		text-decoration: none;
-		list-style: none;
-	}
-}
