+++ client/src/Frames/Header/HeaderFrame.jsx
@@ -1,0 +1,19 @@
+import './HeaderFrame.scss';
+import CurrentTime from '../../components/CurrentTime/CurrentTime';
+import UserNav from '../../components/UserNav/UserNav';
+
+export default function HeaderFrame(props) {
+	const { toggleItems } = props;
+
+	return (
+		<header className='HeaderFrame'>
+			<h2 onClick={() => toggleItems('chat')} className='logo__title'>
+				Listening: 5
+			</h2>
+			<div className='banner'>
+				<CurrentTime twelveHourFormat={true} />
+			</div>
+			<UserNav toggleItems={toggleItems} />
+		</header>
+	);
+}
