+++ client/src/Frames/Information/InformationFrame.scss
@@ -1,0 +1,16 @@
+@use '../../sass/partials/variables' as *;
+
+.InformationFrame {
+	background-color: rgba($color: grey, $alpha: 0.15);
+	position: fixed;
+	bottom: 0;
+	width: 75%;
+	height: 16vh;
+	z-index: 0;
+
+	display: flex;
+	justify-content: space-between;
+	align-items: center;
+
+	padding: 0 1rem;
+}
