+++ client/src/components/Notes/Notes.scss
@@ -1,0 +1,122 @@
+@use '../../sass/partials/variables' as *;
+@use '../../sass/partials/mixins' as *;
+
+.note-section {
+	position: relative;
+	margin: 0 2rem;
+	max-height: 100vh;
+	@include tablet-breakpoint {
+		margin: 0 10%;
+	}
+	&__form {
+		display: flex;
+		flex-direction: column;
+		// max-width: 50rem;
+		margin: 0 auto;
+		padding: 0.5rem 1rem;
+		background-color: $black;
+		border-radius: 0.25rem;
+		opacity: 0.8;
+	}
+
+	&__title {
+		background-color: $black;
+		color: $white;
+		border: none;
+		border-bottom: ridge $white 2px;
+		outline: none;
+		font-size: 1.1rem;
+		padding: 0.3rem 1rem;
+		width: 100%;
+
+		&-container {
+			position: relative;
+		}
+	}
+
+	&__note {
+		background-color: $black;
+		color: $white;
+		outline: none;
+		border: none;
+		resize: none;
+		padding: 1rem;
+		height: 20rem;
+	}
+
+	&__save-note {
+		position: absolute;
+		right: 2rem;
+		text-shadow: $text-shadow-white;
+
+		&:hover {
+			scale: 1.1;
+			color: green;
+		}
+	}
+
+	&__show-archive {
+		position: absolute;
+		z-index: 1;
+		right: 1rem;
+		top: 0.5rem;
+		text-shadow: $text-shadow-white;
+	}
+
+	.fa-archive {
+		&:hover {
+			scale: 1.1;
+			color: green;
+		}
+	}
+
+	.fa-times {
+		&:hover {
+			color: #df1717;
+		}
+	}
+
+	&__history {
+		max-height: 70vh;
+		overflow-y: scroll;
+		background-color: $black;
+		opacity: 0.8;
+	}
+}
+
+.btn {
+	border: none;
+	width: auto;
+	color: $white;
+	background-color: inherit;
+	margin: 0 0.3rem;
+	padding: 0.3rem;
+	width: 1rem;
+
+	i {
+		font-size: 1.1rem;
+	}
+}
+
+.note-card {
+	padding: 1rem 2rem 0.5rem 1rem;
+	position: relative;
+
+	&__title {
+		margin-bottom: 0.5rem;
+	}
+
+	&__timestamp {
+		font-size: 0.8rem;
+		text-shadow: $text-shadow-white;
+	}
+
+	&__delete {
+		font-size: 0.8rem;
+		text-shadow: $text-shadow-white;
+		cursor: pointer;
+		&:hover {
+			color: #df1717;
+		}
+	}
+}
