+++ client/src/components/Settings/Settings.scss
@@ -1,0 +1,24 @@
+@use '../../sass/partials/variables' as *;
+
+.bg-selector {
+	position: absolute;
+	font-size: 1.2rem;
+	right: 0;
+	text-shadow: $text-shadow-white;
+
+	&__icon {
+		cursor: pointer;
+	}
+}
+
+.settings-wrapper {
+	width: 30rem;
+	text-align: right;
+}
+
+.time-button {
+	margin-bottom: 1rem;
+	button {
+		margin-left: 1rem;
+	}
+}
