+++ client/src/components/UserNav/UserNav.scss
@@ -1,0 +1,20 @@
+@use '../../sass/partials/variables' as *;
+
+.user {
+	&__navigation {
+		display: flex;
+		text-shadow: $text-shadow-white;
+	}
+
+	&__icon-container {
+		position: relative;
+		padding: 0.3rem;
+		margin-right: 0.5rem;
+		cursor: pointer;
+	}
+
+	&__icon-title {
+		position: absolute;
+		margin-left: -0.8rem;
+	}
+}
